remove unnecessary MANIFEST.MF additions, relocate the PatternFly palette
