+++ application/org.openjdk.jmc.flightrecorder.ui/META-INF/MANIFEST.MF
@@ -22,4 +22,3 @@ Export-Package: org.openjdk.jmc.flightrecorder.ui,
  org.openjdk.jmc.flightrecorder.ui.preferences,
  org.openjdk.jmc.flightrecorder.ui.selection
 Automatic-Module-Name: org.openjdk.jmc.flightrecorder.ui
-Import-Package: org.eclipse.jdt.core.dom
