+++ application/org.openjdk.jmc.ui.common/META-INF/MANIFEST.MF
@@ -9,9 +9,6 @@ Bundle-Vendor: Oracle Corporation
 Require-Bundle: org.eclipse.core.runtime;visibility:=reexport,
  org.openjdk.jmc.common;visibility:=reexport
 Bundle-ActivationPolicy: lazy
-Import-Package: org.eclipse.core.expressions,
- org.eclipse.swt.graphics,
- org.eclipse.swt.widgets
 Export-Package: org.openjdk.jmc.ui.common,
  org.openjdk.jmc.ui.common.action,
  org.openjdk.jmc.ui.common.idesupport,
